kalman

+++ Drivers/mpu6000.h
@@ -151,9 +151,9 @@ typedef union
 
 typedef struct
 	{
-		int16_t x;
-		int16_t y;
-		int16_t z;
+		float x;
+		float y;
+		float z;
 	}axis;
 	
 enum mpu6000_dlpf_bw {
